+++ CHANGELOG.md
@@ -1093,3 +1093,4 @@ This is a fast forward to v1.6.3 of daviscook477's fork with a few additional ch
 #### dev ####
 * Fix language dropdown going offscreen while in a run (kiooeht)
 * Fix settings toggles being the wrong size while in a run (kiooeht)
+* Fix frame color in SRV being incorrect the first time you view a relic (kiooeht)